@@ -1,63 +1,49 @@
 {
   "name": "suprematism-more-menu",
-  "version": "0.0.1",
+  "version": "0.0.0",
   "license": "MIT",
-  "angular-cli": {},
-  "main": "out-tsc/index.js",
-  "typings": "out-tsc/index.d.ts",
+  "main": "src/index.ts",
+  "typings": "src/index.d.ts",
   "scripts": {
-    "build": "npm run build:tsc && npm run build:css && npm run build:copyAssets",
-    "build:copyAssets": "cp -R src/*.css ./out-tsc && cp -R src/*html ./out-tsc",
-    "build:tsc": "tsc -p src",
-    "build:css": "node-sass src -o src",
-    "deploy": "npm run build && ng github-pages:deploy",
-    "clean": "rm -rf out-tsc && rm -f src/more-menu.component.css",
-    "e2e": "protractor",
-    "example": "npm run watch & ng serve",
-    "lint": "npm run lint:src && npm run lint:example",
-    "lint:src": "tslint \"src/**/*.ts\"",
-    "lint:example": "tslint \"example/**/*.ts\"",
-    "prebuild": "npm run clean",
-    "pree2e": "webdriver-manager update",
-    "prepublish": "npm run build",
-    "postpublish": "npm run clean",
+    "e2e": "ng e2e",
+    "lint": "tslint \"src/**/*.ts\"",
     "test": "ng test",
-    "watch": "npm run watch:css",
-    "watch:css": "node-sass -w src -o src"
+    "start": "ng serve"
   },
   "private": true,
-  "dependencies": {
-    "suprematism-font-icons": "github:cinbcuniversal/suprematism-font-icons",
-    "suprematism-style-guide": "github:cinbcuniversal/suprematism-style-guide"
-  },
+  "dependencies": {},
   "peerDependencies": {
-    "@angular/core": "^2.0.0"
+    "suprematism-style-guide": "~0"
   },
   "devDependencies": {
-    "@angular/common": "2.0.0",
-    "@angular/compiler": "2.0.0",
-    "@angular/core": "2.0.0",
-    "@angular/platform-browser": "2.0.0",
-    "@angular/platform-browser-dynamic": "2.0.0",
-    "@angular/router": "3.0.0",
-    "@types/jasmine": "^2.2.30",
-    "angular-cli": "1.0.0-beta.16",
-    "codelyzer": "~0.0.26",
-    "core-js": "^2.4.1",
-    "jasmine-core": "2.4.1",
-    "jasmine-spec-reporter": "2.5.0",
-    "karma": "1.2.0",
-    "karma-chrome-launcher": "^2.0.0",
-    "karma-cli": "^1.0.1",
-    "karma-jasmine": "^1.0.2",
-    "karma-remap-istanbul": "^0.2.1",
-    "protractor": "4.0.9",
-    "rxjs": "5.0.0-beta.12",
+    "@angular/cli": "1.0.0",
+    "@angular/compiler-cli": "^4.0.0",
+    "@types/jasmine": "2.5.38",
+    "@types/node": "~6.0.60",
+    "codelyzer": "~2.0.0",
+    "jasmine-core": "~2.5.2",
+    "jasmine-spec-reporter": "~3.2.0",
+    "karma": "~1.4.1",
+    "karma-chrome-launcher": "~2.0.0",
+    "karma-cli": "~1.0.1",
+    "karma-jasmine": "~1.1.0",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-coverage-istanbul-reporter": "^0.2.0",
+    "protractor": "~5.1.0",
     "suprematism-style-guide": "github:cinbcuniversal/suprematism-style-guide",
-    "ts-helpers": "^1.1.1",
-    "ts-node": "1.2.1",
-    "tslint": "3.13.0",
-    "typescript": "2.0.2",
-    "zone.js": "^0.6.23"
+    "ts-node": "~2.0.0",
+    "tslint": "~4.5.0",
+    "typescript": "~2.2.0",
+    "@angular/common": "^4.0.0",
+    "@angular/compiler": "^4.0.0",
+    "@angular/core": "^4.0.0",
+    "@angular/forms": "^4.0.0",
+    "@angular/http": "^4.0.0",
+    "@angular/platform-browser": "^4.0.0",
+    "@angular/platform-browser-dynamic": "^4.0.0",
+    "@angular/router": "^4.0.0",
+    "core-js": "^2.4.1",
+    "rxjs": "^5.1.0",
+    "zone.js": "^0.8.4"
   }
 }
